Hipparcos: support sol type 7 and 9 as well.
From Eric Nielsen:
For 5 parameter fits, they’re residuals to the 5 parameter fit. For 7 and 9 parameter fits, they’re residuals to the 5 basic parameter model.
For stochastic/cosmic variance stars, they’re residuals to the 5 parameter fit, but all errors in the IAD were inflated to bring reduced chi-square equal to 1. (So for our fits we deflate the errors by removing that term in quadrature.)
For orbit stars, there’s a secret orbit table for a few dozen stars that always takes me a half hour to find, and the residuals are with respect to that orbit+5 parameter fit.